+++ kymatio/backend/torch_backend.py
@@ -216,3 +216,22 @@ def cdgmm(A, B, inplace=False):
 
 def concatenate(arrays, dim):
     return torch.stack(arrays, dim=dim)
+
+
+def real(x):
+    """Real part of complex tensor
+
+    Takes the real part of a complex tensor, where the last axis corresponds
+    to the real and imaginary parts.
+
+    Parameters
+    ----------
+    x : tensor
+        A complex tensor (that is, whose last dimension is equal to 2).
+
+    Returns
+    -------
+    x_real : tensor
+        The tensor x[..., 0] which is interpreted as the real part of x.
+    """
+    return x[..., 0]
